pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hammerdownloader"
version = "0.1.0"
description = "CLI tool for downloading ride data from Hammerhead API"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
hammerhead = "hammerdownloader.cli:cli"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
